+++ avantis_trader_sdk/__init__.py
@@ -2,4 +2,4 @@ from .client import TraderClient
 from .feed.feed_client import FeedClient
 from .signers.base import BaseSigner
 
-__version__ = "0.8.12"
+__version__ = "0.8.13"